# Conflicts:
#	java/com.sap.sailing.gwt.ui/src/main/java/com/sap/sailing/gwt/ui/server/SailingServiceImpl.java

+++ java/com.sap.sse.security.ui/src/main/java/com/sap/sse/security/ui/server/UserManagementWriteServiceImpl.java
@@ -11,20 +11,14 @@ import java.util.concurrent.Callable;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpSession;
-
 import org.apache.shiro.SecurityUtils;
-import org.apache.shiro.authc.AuthenticationException;
 import org.apache.shiro.authz.AuthorizationException;
 
 import com.sap.sse.common.Util;
 import com.sap.sse.common.Util.Triple;
 import com.sap.sse.common.mail.MailException;
-import com.sap.sse.gwt.client.ServerInfoDTO;
 import com.sap.sse.security.Action;
 import com.sap.sse.security.SecurityService;
-import com.sap.sse.security.interfaces.Credential;
 import com.sap.sse.security.shared.HasPermissions.DefaultActions;
 import com.sap.sse.security.shared.QualifiedObjectIdentifier;
 import com.sap.sse.security.shared.RoleDefinition;
@@ -33,6 +27,8 @@ import com.sap.sse.security.shared.UnauthorizedException;
 import com.sap.sse.security.shared.UserGroupManagementException;
 import com.sap.sse.security.shared.UserManagementException;
 import com.sap.sse.security.shared.WildcardPermission;
+import com.sap.sse.security.shared.dto.AccessControlListDTO;
+import com.sap.sse.security.shared.dto.OwnershipDTO;
 import com.sap.sse.security.shared.dto.RoleDefinitionDTO;
 import com.sap.sse.security.shared.dto.StrippedUserDTO;
 import com.sap.sse.security.shared.dto.StrippedUserGroupDTO;
@@ -55,7 +51,19 @@ import com.sap.sse.security.ui.shared.SuccessInfo;
 public class UserManagementWriteServiceImpl extends UserManagementServiceImpl implements UserManagementWriteService {
     private static final long serialVersionUID = -8123229851467370537L;
     private static final Logger logger = Logger.getLogger(UserManagementWriteServiceImpl.class.getName());
-    
+
+    @Override
+    public OwnershipDTO setOwnership(final String username, final UUID userGroupId,
+            final QualifiedObjectIdentifier idOfOwnedObject, final String displayNameOfOwnedObject) {
+        SecurityUtils.getSubject()
+                .checkPermission(idOfOwnedObject.getStringPermission(DefaultActions.CHANGE_OWNERSHIP));
+        final User user = getSecurityService().getUserByName(username);
+        // no security check if current user can see the user associated with the given username
+        final Ownership result = getSecurityService().setOwnership(idOfOwnedObject, user,
+                getSecurityService().getUserGroup(userGroupId), displayNameOfOwnedObject);
+        return securityDTOFactory.createOwnershipDTO(result, new HashMap<>(), new HashMap<>());
+    }
+
     @Override
     public RoleDefinitionDTO createRoleDefinition(String roleDefinitionIdAsString, String name) {
         RoleDefinition role = getSecurityService().setOwnershipWithoutCheckPermissionForObjectCreationAndRevertOnError(
@@ -105,27 +113,22 @@ public class UserManagementWriteServiceImpl extends UserManagementServiceImpl im
     public void updateRoleDefinition(RoleDefinitionDTO roleDefinitionWithNewProperties) throws UnauthorizedException {
         SecurityUtils.getSubject().checkPermission(SecuredSecurityTypes.ROLE_DEFINITION.getStringPermissionForObject(
                 DefaultActions.UPDATE, roleDefinitionWithNewProperties));
-        
         RoleDefinition existingRole = getSecurityService().getRoleDefinition(roleDefinitionWithNewProperties.getId());
         if (existingRole == null) {
             throw new UnauthorizedException("Role does not exist");
         }
         Set<WildcardPermission> addedPermissions = new HashSet<>(roleDefinitionWithNewProperties.getPermissions());
         addedPermissions.removeAll(existingRole.getPermissions());
-        
         if (!getSecurityService().hasUserAllWildcardPermissionsForAlreadyRealizedQualifications(existingRole, addedPermissions)) {
             throw new UnauthorizedException("Not permitted to grant permissions for role "
                     + roleDefinitionWithNewProperties.getName());
         }
-        
         Set<WildcardPermission> removedPermissions = new HashSet<>(existingRole.getPermissions());
         removedPermissions.removeAll(roleDefinitionWithNewProperties.getPermissions());
-        
         if (!getSecurityService().hasUserAllWildcardPermissionsForAlreadyRealizedQualifications(existingRole, removedPermissions)) {
             throw new UnauthorizedException("Not permitted to revoke permissions for role "
                     + roleDefinitionWithNewProperties.getName());
         }
-        
         getSecurityService().updateRoleDefinition(roleDefinitionWithNewProperties);
     }
 
@@ -194,10 +197,18 @@ public class UserManagementWriteServiceImpl extends UserManagementServiceImpl im
     }
 
     @Override
-    public void removeUserFromUserGroup(String userGroupIdAsString, String username) throws UnauthorizedException {
+    public void removeUserFromUserGroup(String userGroupIdAsString, String username)
+            throws UnauthorizedException {
         final UserGroup userGroup = getSecurityService().getUserGroup(UUID.fromString(userGroupIdAsString));
-        if (SecurityUtils.getSubject().isPermitted(SecuredSecurityTypes.USER_GROUP.getStringPermissionForObject(DefaultActions.DELETE, userGroup))) {
-            getSecurityService().removeUserFromUserGroup(userGroup, getSecurityService().getUserByName(username));
+        if (SecurityUtils.getSubject().isPermitted(
+                SecuredSecurityTypes.USER_GROUP.getStringPermissionForObject(DefaultActions.UPDATE, userGroup))) {
+            if (getSecurityService().hasCurrentUserMetaPermissionsOfRoleDefinitionsWithQualification(
+                    userGroup.getRoleDefinitionMap().keySet(), new Ownership(null, userGroup))) {
+                getSecurityService().removeUserFromUserGroup(userGroup, getSecurityService().getUserByName(username));
+            } else {
+                throw new UnauthorizedException(
+                        "Current user does not have all the meta permissions of the user group the user would be removed from");
+            }
         } else {
             throw new UnauthorizedException("Not permitted to remove user from group");
         }
@@ -232,36 +243,14 @@ public class UserManagementWriteServiceImpl extends UserManagementServiceImpl im
             final RoleDefinition roleDefinition = getSecurityService()
                     .getRoleDefinition(UUID.fromString(roleDefinitionIdAsString));
             if (roleDefinition != null) {
+                if (!getSecurityService().hasCurrentUserMetaPermissionsOfRoleDefinitionWithQualification(roleDefinition,
+                        new Ownership(null, userGroup))) {
+                    throw new UnauthorizedException("Not permitted to remove role definition from group");
+                }
                 getSecurityService().removeRoleDefintionFromUserGroup(userGroup, roleDefinition);
             }
         } else {
             throw new UnauthorizedException("Not permitted to remove role definition from group");
-        }
-    }
-    
-    @Override
-    public SuccessInfo logout() {
-        logger.info("Logging out user: " + SecurityUtils.getSubject());
-        getSecurityService().logout();
-        getHttpSession().invalidate();
-        final Cookie cookie = new Cookie(UserManagementConstants.LOCALE_COOKIE_NAME, "");
-        cookie.setMaxAge(0);
-        cookie.setPath("/");
-        getThreadLocalResponse().addCookie(cookie);
-        logger.info("Invalidated HTTP session");
-        return new SuccessInfo(true, "Logged out.", /* redirectURL */ null, null);
-    }
-
-    @Override
-    public SuccessInfo login(String username, String password) {
-        try {
-            String redirectURL = getSecurityService().login(username, password);
-            UserDTO user = securityDTOFactory.createUserDTOFromUser(getSecurityService().getUserByName(username),
-                    getSecurityService());
-            return new SuccessInfo(true, "Success. Redirecting to " + redirectURL, redirectURL,
-                    new Triple<>(user, getAllUser(), getServerInfo()));
-        } catch (UserManagementException | AuthenticationException e) {
-            return new SuccessInfo(false, SuccessInfo.FAILED_TO_LOGIN, /* redirectURL */ null, null);
         }
     }
 
@@ -402,18 +391,6 @@ public class UserManagementWriteServiceImpl extends UserManagementServiceImpl im
     }
 
     @Override
-    public Triple<UserDTO, UserDTO, ServerInfoDTO> verifySocialUser(CredentialDTO credentialDTO) {
-        User user = null;
-        try {
-            user = getSecurityService().verifySocialUser(createCredentialFromDTO(credentialDTO));
-        } catch (UserManagementException e) {
-            e.printStackTrace();
-        }
-        final UserDTO userDTO = securityDTOFactory.createUserDTOFromUser(user, getSecurityService());
-        return new Triple<>(userDTO, getAllUser(), getServerInfo());
-    }
-
-    @Override
     public void addSetting(String key, String clazz, String setting) {
         try {
             getSecurityService().addSetting(key, Class.forName(clazz));
@@ -463,18 +440,13 @@ public class UserManagementWriteServiceImpl extends UserManagementServiceImpl im
         try {
             // get user for which to add a role
             final User user = getOrThrowUser(username);
-
             // get user for which the role is qualified, if one exists
             getOrThrowQualifiedUser(userQualifierName);
-
             // get the group tenant the role is qualified for if one exists
             final UserGroup tenant = getOrThrowTenant(tenantQualifierName);
-
-            final Role role = getOrThrowRoleFromIDs(roleDefinitionId, tenant == null ? null : tenant.getId(),
+            final Role role = getOrThrowRoleFromIDsAndCheckMetaPermissions(roleDefinitionId, tenant == null ? null : tenant.getId(),
                     userQualifierName);
-
             final TypeRelativeObjectIdentifier associationTypeIdentifier = PermissionAndRoleAssociation.get(role, user);
-
             final String message = "added role " + role.getName() + " for user " + username;
             getSecurityService().setOwnershipWithoutCheckPermissionForObjectCreationAndRevertOnError(
                     SecuredSecurityTypes.ROLE_ASSOCIATION, associationTypeIdentifier,
@@ -512,7 +484,7 @@ public class UserManagementWriteServiceImpl extends UserManagementServiceImpl im
             getOrThrowQualifiedUser(userQualifierName);
             // get the group tenant the role is qualified for if one exists
             UserGroup tenant = getOrThrowTenant(tenantQualifierName);
-            Role role = getOrThrowRoleFromIDs(roleDefinitionId, tenant == null ? null : tenant.getId(),
+            Role role = getOrThrowRoleFromIDsAndCheckMetaPermissions(roleDefinitionId, tenant == null ? null : tenant.getId(),
                     userQualifierName);
             final String message = "removed role " + role.getName() + " for user " + username;
             final TypeRelativeObjectIdentifier associationTypeIdentifier = PermissionAndRoleAssociation.get(role, user);
@@ -533,7 +505,7 @@ public class UserManagementWriteServiceImpl extends UserManagementServiceImpl im
         } catch (UserManagementException e) {
             successInfo = new SuccessInfo(false,
                     "You are not allowed to revoke this role from user " + username
-                            + " or the username, grou name or role name did not exist.",
+                            + " or the username, group name or role name did not exist.",
                     /* redirectURL */ null, /* userDTO */ null);
         }
         return successInfo;
@@ -552,8 +524,7 @@ public class UserManagementWriteServiceImpl extends UserManagementServiceImpl im
                         "Not permitted to grant/revoke permission " + permission + " for user " + user.getName());
             }
             // grant permission
-            final TypeRelativeObjectIdentifier associationTypeIdentifier = PermissionAndRoleAssociation.get(permission,
-                    user);
+            final TypeRelativeObjectIdentifier associationTypeIdentifier = PermissionAndRoleAssociation.get(permission, user);
             final String message = "Added permission " + permission + " for user " + username;
             getSecurityService().setOwnershipWithoutCheckPermissionForObjectCreationAndRevertOnError(
                     SecuredSecurityTypes.PERMISSION_ASSOCIATION, associationTypeIdentifier,
@@ -585,6 +556,11 @@ public class UserManagementWriteServiceImpl extends UserManagementServiceImpl im
         try {
             // check if user exists
             User user = getOrThrowUser(username);
+            // check permissions
+            if (!getSecurityService().hasCurrentUserMetaPermissionWithOwnershipLookup(permission)) {
+                throw new UnauthorizedException(
+                        "Not permitted to grant/revoke permission " + permission + " for user " + user.getName());
+            }
             // revoke permission
             final String message = "Revoked permission " + permission + " for user " + username;
             final TypeRelativeObjectIdentifier associationTypeIdentifier = PermissionAndRoleAssociation.get(permission,
@@ -610,20 +586,6 @@ public class UserManagementWriteServiceImpl extends UserManagementServiceImpl im
         return successInfo;
     }
 
-    protected Credential createCredentialFromDTO(CredentialDTO credentialDTO) {
-        Credential credential = new Credential();
-        credential.setAuthProvider(credentialDTO.getAuthProvider());
-        credential.setAuthProviderName(credentialDTO.getAuthProviderName());
-        credential.setEmail(credentialDTO.getEmail());
-        credential.setLoginName(credentialDTO.getLoginName());
-        credential.setPassword(credentialDTO.getPassword());
-        credential.setRedirectUrl(credentialDTO.getRedirectUrl());
-        credential.setState(credentialDTO.getState());
-        credential.setVerifier(credentialDTO.getVerifier());
-        credential.setOauthToken(credentialDTO.getOauthToken());
-        return credential;
-    }
-
     protected void sendPasswordChangedMailAsync(final String username) {
         new Thread("sending updated password to user "+username+" by e-mail") {
             @Override public void run() {
@@ -645,10 +607,6 @@ public class UserManagementWriteServiceImpl extends UserManagementServiceImpl im
         }
     }
 
-    protected HttpSession getHttpSession() {
-        return getThreadLocalRequest().getSession();
-    }
-
     /**
      * @returns the user associated with the userQualifierName or null
      * @throws UserManagementException
@@ -668,7 +626,7 @@ public class UserManagementWriteServiceImpl extends UserManagementServiceImpl im
      *             if the current user does not have the meta permission to give this specific, qualified role in this
      *             context.
      */
-    protected Role getOrThrowRoleFromIDs(UUID roleDefinitionId, UUID tenantId, String userQualifierName) throws UserManagementException {
+    protected Role getOrThrowRoleFromIDsAndCheckMetaPermissions(UUID roleDefinitionId, UUID tenantId, String userQualifierName) throws UserManagementException {
         final Role role = createRoleFromIDs(roleDefinitionId, tenantId, userQualifierName);
         if (!getSecurityService().hasCurrentUserMetaPermissionsOfRoleDefinitionWithQualification(
                 role.getRoleDefinition(), role.getQualificationAsOwnership())) {
@@ -720,6 +678,31 @@ public class UserManagementWriteServiceImpl extends UserManagementServiceImpl im
         return new Role(
                 getSecurityService().getRoleDefinition(roleDefinitionId),
                 qualifyingTenantId == null ? null : getSecurityService().getUserGroup(qualifyingTenantId), user);
+    }
+
+    @Override
+    public AccessControlListDTO overrideAccessControlList(QualifiedObjectIdentifier idOfAccessControlledObject,
+            AccessControlListDTO acl) throws UnauthorizedException {
+        if (SecurityUtils.getSubject()
+                .isPermitted(idOfAccessControlledObject.getStringPermission(DefaultActions.CHANGE_ACL))) {
+            
+            Map<UserGroup, Set<String>> aclActionsByGroup = new HashMap<>();
+            for (Entry<StrippedUserGroupDTO, Set<String>> entry : acl.getActionsByUserGroup().entrySet()) {
+                final StrippedUserGroupDTO groupDTO = entry.getKey();
+                final UserGroup userGroup;
+                if (groupDTO == null) {
+                    userGroup = null;
+                } else {
+                    userGroup = getSecurityService().getUserGroup(groupDTO.getId());
+                }
+                aclActionsByGroup.put(userGroup, entry.getValue());
+            }
+
+            return securityDTOFactory.createAccessControlListDTO(getSecurityService()
+                    .overrideAccessControlList(idOfAccessControlledObject, aclActionsByGroup));
+        } else {
+            throw new UnauthorizedException("Not permitted to update the ACL for a user");
+        }
     }
 
 }
